Gizmo fix
Gizmos only seem to work in Unity if there is a camera in the scene.

Due to the preload scene carrying the desired main camera through to every other scene, this presents a problem. Multiple cameras mean multiple audio handlers, multiple video feeds would harm performance, etc.

To resolve this I added a simple script to the main camera prefab that causes it to destroy itself on awake with a boolean check that can be set in the editor. The preload scene has this flag set to false, the main camera has it set to true by default. This means other developers can add the main camera without worrying about having multiple in the scene at the same time.

I also added some checks in to the various visualizers to eliminate erroneous error messages.